+++ packages/core/src/model/colorpicker/dynamic-colorpicker.model.ts
@@ -1,5 +1,5 @@
 import { DynamicFormValueControlModelConfig, DynamicFormValueControlModel } from "../dynamic-form-value-control.model";
-import { DynamicFormControlLayout } from "../dynamic-form-control-layout.model";
+import { DynamicFormControlLayout } from "../misc/dynamic-form-control-layout.model";
 import { serializable } from "../../decorator/serializable.decorator";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_COLORPICKER = "COLORPICKER";