+++ packages/core/src/model/datepicker/dynamic-datepicker.model.ts
@@ -1,10 +1,10 @@
-import { DynamicFormControlLayout } from "../dynamic-form-control-layout.model";
-import { serializable } from "../../decorator/serializable.decorator";
 import {
     DynamicDateControlModel,
     DynamicDateControlModelConfig,
     DynamicDateControlValue
 } from "../dynamic-date-control.model";
+import { DynamicFormControlLayout } from "../misc/dynamic-form-control-layout.model";
+import { serializable } from "../../decorator/serializable.decorator";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER = "DATEPICKER";
 
@@ -13,6 +13,7 @@ export interface DynamicDatePickerModelConfig extends DynamicDateControlModelCon
     focusedDate?: DynamicDateControlValue;
     inline?: boolean;
     toggleIcon?: string;
+    toggleLabel?: string;
 }
 
 export class DynamicDatePickerModel extends DynamicDateControlModel {
@@ -20,6 +21,7 @@ export class DynamicDatePickerModel extends DynamicDateControlModel {
     @serializable() focusedDate: DynamicDateControlValue | null;
     @serializable() inline: boolean;
     @serializable() toggleIcon: string | null;
+    @serializable() toggleLabel: string | null;
 
     @serializable() readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER;
 
@@ -30,5 +32,6 @@ export class DynamicDatePickerModel extends DynamicDateControlModel {
         this.focusedDate = config.focusedDate || null;
         this.inline = typeof config.inline === "boolean" ? config.inline : false;
         this.toggleIcon = typeof config.toggleIcon === "string" ? config.toggleIcon : null;
+        this.toggleLabel = typeof config.toggleLabel === "string" ? config.toggleLabel : null;
     }
 }