+++ packages/core/src/model/dynamic-form-value-control.model.ts
@@ -1,6 +1,6 @@
 import { Subject } from "rxjs/Subject";
 import { DynamicFormControlModel, DynamicFormControlModelConfig } from "./dynamic-form-control.model";
-import { DynamicFormControlLayout } from "./dynamic-form-control-layout.model";
+import { DynamicFormControlLayout } from "./misc/dynamic-form-control-layout.model";
 import { serializable } from "../decorator/serializable.decorator";
 
 export type DynamicFormControlValue = boolean | number | string | object | Date |