+++ packages/core/src/model/input/dynamic-input.model.ts
@@ -1,5 +1,5 @@
 import { DynamicInputControlModel, DynamicInputControlModelConfig } from "../dynamic-input-control.model";
-import { DynamicFormControlLayout } from "../dynamic-form-control-layout.model";
+import { DynamicFormControlLayout } from "../misc/dynamic-form-control-layout.model";
 import { serializable } from "../../decorator/serializable.decorator";
 import { Utils } from "../../utils/core.utils";