+++ packages/core/src/model/misc/dynamic-form-control-path.model.ts
@@ -1,0 +1,7 @@
+export interface DynamicPathable {
+
+    id?: string;
+    index?: number | null;
+    parent: DynamicPathable | null;
+}
+