+++ packages/core/src/model/misc/dynamic-form-control-validation.model.ts
@@ -1,0 +1,7 @@
+export interface DynamicValidatorDescriptor {
+
+    name: string;
+    args: any;
+}
+
+export type DynamicValidatorsConfig = { [validatorKey: string]: any | DynamicValidatorDescriptor };