+++ packages/core/src/model/select/dynamic-select.model.ts
@@ -1,5 +1,5 @@
 import { DynamicOptionControlModel, DynamicOptionControlModelConfig } from "../dynamic-option-control.model";
-import { DynamicFormControlLayout } from "../dynamic-form-control-layout.model";
+import { DynamicFormControlLayout } from "../misc/dynamic-form-control-layout.model";
 import { serializable } from "../../decorator/serializable.decorator";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_SELECT = "SELECT";