+++ packages/ui-bootstrap/src/dynamic-bootstrap-form-ui.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { ReactiveFormsModule } from "@angular/forms";
-//import { BsDatepickerModule } from "ngx-bootstrap";
-//import { TimepickerModule } from "ngx-bootstrap";
+import { BsDatepickerModule } from "ngx-bootstrap/datepicker";
+import { TimepickerModule } from "ngx-bootstrap/timepicker";
 import { TextMaskModule } from "angular2-text-mask";
 import { DynamicFormsCoreModule } from "@ng-dynamic-forms/core";
 import { DynamicBootstrapFormControlComponent } from "./dynamic-bootstrap-form-control.component";
@@ -12,8 +12,8 @@ import { DynamicBootstrapFormComponent } from "./dynamic-bootstrap-form.componen
     imports: [
         CommonModule,
         ReactiveFormsModule,
-        //BsDatepickerModule,
-        //TimePickerModule.forRoot()
+        BsDatepickerModule.forRoot(),
+        TimepickerModule.forRoot(),
         TextMaskModule,
         DynamicFormsCoreModule
     ],