+++ sample/app/ui-material/material-sample-form.model.ts
@@ -109,7 +109,7 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         errorMessages: {
             required: "Field is required"
         },
-        custom: {
+        additional: {
             color: "accent"
         }
     }),